requirements.txt

streamlit
pandas
numpy
scikit-learn
joblib
tensorflow-cpu
